requirements.txt
langchain 
langchain-community 
pypdf 
faiss-cpu
streamlit 
python-dotenv